requirements.txt
absl-py==1.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcolz==1.2.1
beautifulsoup4==4.11.1
black==22.1.0
bleach==4.1.0
bokeh @ file:///C:/ci/bokeh_1638362940011/work
Bottleneck @ file:///C:/ci/bottleneck_1648029103672/work
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click @ file:///C:/ci/click_1646056799533/work
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.11.0
Cython @ file:///C:/ci/cython_1647852217814/work
cytoolz==0.11.0
dask @ file:///opt/conda/conda-bld/dask-core_1647268715755/work
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distributed @ file:///opt/conda/conda-bld/distributed_1647271944416/work
Django==4.0.4
entrypoints==0.4
executing==0.8.2
fonttools==4.29.1
fsspec @ file:///opt/conda/conda-bld/fsspec_1647268051896/work
gast==0.3.3
gensim==4.1.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==2.10.0
HeapDict @ file:///Users/ktietz/demo/mc3/conda-bld/heapdict_1630598515714/work
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1643987515720/work/keras-2.8.0-py2.py3-none-any.whl
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
locket @ file:///C:/ci/locket_1647006279754/work
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
mock @ file:///tmp/build/80754af9/mock_1607622725907/work
msgpack @ file:///C:/ci/msgpack-python_1612287368835/work
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.7
notebook==6.4.8
numexpr @ file:///C:/ci/numexpr_1640704408721/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1649782971440/work
oauthlib==3.2.0
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1647955046463/work
pandocfilters==1.5.0
parso==0.8.3
partd @ file:///opt/conda/conda-bld/partd_1647245470509/work
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil @ file:///C:/ci/psutil_1612298324802/work
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pywin32==303
pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.4.1
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==5.2.1
sortedcontainers @ file:///tmp/build/80754af9/sortedcontainers_1623949099177/work
soupsieve==2.3.2
sqlparse==0.4.2
stack-data==0.2.0
tables==3.7.0
tblib @ file:///Users/ktietz/demo/mc3/conda-bld/tblib_1629402031467/work
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
tomli==2.0.1
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
torch==1.11.0
torchvision==0.12.0
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm==4.62.3
traitlets==5.1.1
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
tzdata==2022.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wincertstore==0.2
wrapt==1.13.3
zict==2.0.0
zipp==3.7.0
